refactor(shamhub): Split CLI admin client
The ShamHub CLI keeps command dispatch, command implementations, and its
admin REST transport in one package. Move the transport client into
cli_admin_client.go so changes to endpoint communication are local to the
client file while cli_main.go stays centered on command dispatch and command
behavior.

This is a package-local file organization change. Command syntax,
environment variables, endpoint paths, request bodies, and response handling
remain unchanged.
